Move multiparty ecdh tests.. Also typehint decrypt() against EncryptedMessage
